+++ src/test/java/nl/knaw/dans/dvcli/AbstractCapturingTest.java
@@ -19,13 +19,35 @@ import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.read.ListAppender;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.slf4j.LoggerFactory;
 
 import java.io.ByteArrayOutputStream;
+import java.io.OutputStream;
 import java.io.PrintStream;
-import java.util.stream.Stream;
 
-public class TestUtils {
+public abstract class AbstractCapturingTest {
+    private final PrintStream originalStdout = System.out;
+    private final PrintStream originalStderr = System.err;
+    protected OutputStream stdout;
+    protected OutputStream stderr;
+    protected ListAppender<ILoggingEvent> logged;
+
+    @AfterEach
+    public void tearDown() {
+
+        System.setOut(originalStdout);
+        System.setErr(originalStderr);
+    }
+
+    @BeforeEach
+    public void setUp() {
+        stdout = captureStdout();
+        stderr = captureStderr();
+        logged = captureLog(Level.DEBUG, "nl.knaw.dans");
+    }
+
     public static ListAppender<ILoggingEvent> captureLog(Level error, String loggerName) {
         var logger = (Logger) LoggerFactory.getLogger(loggerName);
         ListAppender<ILoggingEvent> listAppender = new ListAppender<>();
@@ -45,9 +67,5 @@ public class TestUtils {
         var outContent = new ByteArrayOutputStream();
         System.setErr(new PrintStream(outContent));
         return outContent;
-    }
-
-    public static Stream<String> messagesOf(ListAppender<ILoggingEvent> logged) {
-        return logged.list.stream().map(iLoggingEvent -> iLoggingEvent.getLevel() + "  " + iLoggingEvent.getFormattedMessage());
     }
 }
