+++ src/test/java/nl/knaw/dans/dvcli/AbstractTestWithTestDir.java
@@ -26,7 +26,6 @@ public abstract class AbstractTestWithTestDir {
     protected final Path testDir = Path.of("target/test")
         .resolve(getClass().getSimpleName());
 
-
     @BeforeEach
     public void setUp() throws Exception {
         if (testDir.toFile().exists()) {
